Move reservationUtils to own file from dataUtils
Refs #397.
