+++ app/components/reservation/ConfirmReservationModal.js
@@ -6,7 +6,7 @@ import Modal from 'react-bootstrap/lib/Modal';
 import CompactReservationsList from 'components/common/CompactReservationsList';
 import constants from 'constants/AppConstants';
 import ReservationForm from 'containers/ReservationForm';
-import { isStaffEvent } from 'utils/DataUtils';
+import { isStaffEvent } from 'utils/reservationUtils';
 
 class ConfirmReservationModal extends Component {
   constructor(props) {
